requirements.txt
flask==2.3.3
selenium==4.15.2
webdriver-manager==4.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
pdf2image==1.16.3
pillow==10.0.1
numpy==1.25.2
requests==2.31.0